Reduce headers needed by election.hpp and moves election_status in to its own file
Before:
[39/39] Linking CXX executable rpc_test
ninja  338.18s user 25.74s system 714% cpu 50.953 total
After:
[39/39] Linking CXX executable rpc_test
ninja  339.33s user 27.66s system 778% cpu 47.155 total
